+++ database/migrations/2025_09_23_221050_create_mission_reward_competences_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('mission_reward_competences', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedInteger('mission_id')->comment('Идентификатор миссии');
+            $table->unsignedBigInteger('mission_id')->comment('Идентификатор миссии');
             $table->foreign('mission_id')->references('id')->on('missions')->index('mission_id')->onDelete('cascade')->comment('Идентификатор миссии');
 
-            $table->unsignedInteger('competence_id')->comment('Идентификатор компетенции');
+            $table->unsignedBigInteger('competence_id')->comment('Идентификатор компетенции');
             $table->foreign('competence_id')->references('id')->on('сompetencies')->index('competence_id')->onDelete('cascade')->comment('Идентификатор компетенции');
 
             $table->integer('level_increase')->comment('Увеличение уровня компетенции за выполнение миссии');
