+++ database/migrations/2025_09_25_000216_create_mission_requirements_table.php
@@ -14,13 +14,13 @@ return new class extends Migration
         Schema::create('mission_requirements', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedInteger('mission_id')->comment('Идентификатор миссии');
+            $table->unsignedBigInteger('mission_id')->comment('Идентификатор миссии');
             $table->foreign('mission_id')->references('id')->on('missions')->index('mission_id')->onDelete('cascade');
 
-            $table->unsignedInteger('branch_requirement_id')->nullable()->comment('Выполненная ветка миссий, которая требуется для активации миссии');
+            $table->unsignedBigInteger('branch_requirement_id')->nullable()->comment('Выполненная ветка миссий, которая требуется для активации миссии');
             $table->foreign('branch_requirement_id')->references('id')->on('branches')->index('branch_requirement_id')->onDelete('cascade');
 
-            $table->unsignedInteger('mission_requirement_id')->nullable()->comment('Выполненная миссия, которая требуется для активации миссии');
+            $table->unsignedBigInteger('mission_requirement_id')->nullable()->comment('Выполненная миссия, которая требуется для активации миссии');
             $table->foreign('mission_requirement_id')->references('id')->on('missions')->index('mission_requirement_id')->onDelete('cascade');
 
             $table->timestamps();
